Finalize proposal for select()
