+++ build.gradle
@@ -23,5 +23,8 @@ dependencies {
     testCompile 'org.postgresql:postgresql:9.4.1208.jre7'
     testCompile 'junit:junit:4.+',
                 'mysql:mysql-connector-java:5.1.37',
-                'com.h2database:h2:1.4.190'
+                'com.h2database:h2:1.4.190',
+                'org.xerial:sqlite-jdbc:3.8.+',
+                'org.jetbrains.kotlin:kotlin-test:1.0.+'
+
 }
