--- src/main/kotlin/kdbc/ConnectionProvider.kt
@@ -1,9 +1,0 @@
-package kdbc
-
-import java.sql.Connection
-
-interface ConnectionProvider {
-
-    fun get(): Connection
-    fun close(): Unit
-}
